FEAT: Refactor setup.py with sub-modules
