+++ .github/workflows/build.yml
@@ -17,6 +17,7 @@ concurrency:
 env:
   COINCURVE_UPSTREAM_REF: 1ad5185cd42c0636104129fcc9f6a4bf9c67cc40
   COINCURVE_IGNORE_SYSTEM_LIB: '1'
+  CIBW_PROJECT_REQUIRES_PYTHON: ">=3.8"
   CIBW_BEFORE_ALL_MACOS: ./.github/scripts/install-macos-build-deps.sh
   CIBW_ENVIRONMENT_PASS_LINUX: >
     COINCURVE_UPSTREAM_REF
