+++ MANIFEST.in
@@ -1,5 +1,5 @@
 include src/coincurve/py.typed
-include setup_support.py
+recursive-include setup_tools *.py
 recursive-include _cffi_build *.py *.h
 prune /*
 graft libsecp256k1
