+++ setup_tools/build_clib.py
@@ -1,0 +1,121 @@
+import errno
+import logging
+import os
+import subprocess
+
+from setuptools.command import build_clib
+
+
+class BuildClib(build_clib.build_clib):
+    def initialize_options(self):
+        super().initialize_options()
+        self.build_flags = None
+
+    def finalize_options(self):
+        super().finalize_options()
+        if self.build_flags is None:
+            self.build_flags = {'include_dirs': [], 'library_dirs': [], 'define': []}
+
+    def get_source_files(self):
+        from support import absolute, has_system_lib
+
+        # Ensure library has been downloaded (sdist might have been skipped)
+        if not has_system_lib():
+            from setup_tools.support import download_library
+
+            download_library(self)
+
+        return [
+            absolute(os.path.join(root, filename))
+            for root, _, filenames in os.walk(absolute('libsecp256k1'))
+            for filename in filenames
+        ]
+
+    def build_libraries(self, libraries):
+        raise Exception('build_libraries')
+
+    def check_library_list(self, libraries):
+        raise Exception('check_library_list')
+
+    def get_library_names(self):
+        from support import build_flags
+
+        return build_flags('libsecp256k1', 'l', os.path.abspath(self.build_temp))
+
+    def run(self):
+        from support import absolute, build_flags, has_system_lib
+
+        from setup import MAKE
+
+        if has_system_lib():
+            logging.info('Using system library')
+            return
+
+        build_temp = os.path.abspath(self.build_temp)
+
+        try:
+            os.makedirs(build_temp)
+        except OSError as e:
+            if e.errno != errno.EEXIST:
+                raise
+
+        if not os.path.exists(absolute('libsecp256k1')):
+            # library needs to be downloaded
+            self.get_source_files()
+
+        if not os.path.exists(absolute('libsecp256k1/configure')):
+            # configure script hasn't been generated yet
+            autogen = absolute('libsecp256k1/autogen.sh')
+            os.chmod(absolute(autogen), 0o700)
+            subprocess.check_call([autogen], cwd=absolute('libsecp256k1'))  # noqa S603
+
+        for filename in [
+            'libsecp256k1/configure',
+            'libsecp256k1/build-aux/compile',
+            'libsecp256k1/build-aux/config.guess',
+            'libsecp256k1/build-aux/config.sub',
+            'libsecp256k1/build-aux/depcomp',
+            'libsecp256k1/build-aux/install-sh',
+            'libsecp256k1/build-aux/missing',
+            'libsecp256k1/build-aux/test-driver',
+        ]:
+            try:
+                os.chmod(absolute(filename), 0o700)
+            except OSError as e:
+                # some of these files might not exist depending on autoconf version
+                if e.errno != errno.ENOENT:
+                    # If the error isn't 'No such file or directory' something
+                    # else is wrong and we want to know about it
+                    raise
+
+        cmd = [
+            absolute('libsecp256k1/configure'),
+            '--disable-shared',
+            '--enable-static',
+            '--disable-dependency-tracking',
+            '--with-pic',
+            '--enable-module-extrakeys',
+            '--enable-module-recovery',
+            '--enable-module-schnorrsig',
+            '--prefix',
+            os.path.abspath(self.build_clib),
+            '--enable-experimental',
+            '--enable-module-ecdh',
+            '--enable-benchmark=no',
+            '--enable-tests=no',
+            '--enable-exhaustive-tests=no',
+        ]
+        if 'COINCURVE_CROSS_HOST' in os.environ:
+            cmd.append(f"--host={os.environ['COINCURVE_CROSS_HOST']}")
+
+        logging.debug(f"Running configure: {' '.join(cmd)}")
+        subprocess.check_call(cmd, cwd=build_temp)  # noqa S603
+
+        subprocess.check_call([MAKE], cwd=build_temp)  # noqa S603
+        subprocess.check_call([MAKE, 'check'], cwd=build_temp)  # noqa S603
+        subprocess.check_call([MAKE, 'install'], cwd=build_temp)  # noqa S603
+
+        self.build_flags['include_dirs'].extend(build_flags('libsecp256k1', 'I', build_temp))
+        self.build_flags['library_dirs'].extend(build_flags('libsecp256k1', 'L', build_temp))
+        if not has_system_lib():
+            self.build_flags['define'].append(('CFFI_ENABLE_RECOVERY', None))
